requirements.txt
Flask==2.0.1
Pillow==9.5.0
pydub==0.25.1
Werkzeug==2.0.3
gunicorn==20.1.0
cloudconvert==2.0.0
python-dotenv==0.19.0 